+++ src/chain_sync/chain_muxer.rs
@@ -152,6 +152,9 @@ pub struct ChainMuxer<DB, M> {
 
     /// Syncing configurations
     sync_config: SyncConfig,
+
+    /// When `stateless_mode` is true, forest connects to the P2P network but does not sync to HEAD.
+    stateless_mode: bool,
 }
 
 impl<DB, M> ChainMuxer<DB, M>
@@ -170,6 +173,7 @@ where
         tipset_sender: flume::Sender<Arc<Tipset>>,
         tipset_receiver: flume::Receiver<Arc<Tipset>>,
         cfg: SyncConfig,
+        stateless_mode: bool,
     ) -> Result<Self, ChainMuxerError> {
         let network =
             SyncNetworkContext::new(network_send, peer_manager, state_manager.blockstore_owned());
@@ -186,6 +190,7 @@ where
             tipset_sender,
             tipset_receiver,
             sync_config: cfg,
+            stateless_mode,
         })
     }
 
@@ -247,10 +252,10 @@ where
         peer_id: PeerId,
         genesis_block_cid: Cid,
     ) {
-        // Query the heaviest TipSet from the store
-        let heaviest = chain_store.heaviest_tipset();
         if network.peer_manager().is_peer_new(&peer_id).await {
             // Since the peer is new, send them a hello request
+            // Query the heaviest TipSet from the store
+            let heaviest = chain_store.heaviest_tipset();
             let request = HelloRequest {
                 heaviest_tip_set: heaviest.cids(),
                 heaviest_tipset_height: heaviest.epoch(),
@@ -497,9 +502,7 @@ where
 
         // Store block messages in the block store
         for block in tipset.blocks() {
-            crate::chain::persist_objects(&chain_store.db, &[block.header()])?;
-            crate::chain::persist_objects(&chain_store.db, block.bls_msgs())?;
-            crate::chain::persist_objects(&chain_store.db, block.secp_msgs())?;
+            block.persist(&chain_store.db)?;
         }
 
         // Update the peer head
@@ -512,6 +515,48 @@ where
             .set(tipset.epoch());
 
         Ok(Some((tipset, source)))
+    }
+
+    fn stateless_node(&self) -> ChainMuxerFuture<(), ChainMuxerError> {
+        let p2p_messages = self.net_handler.clone();
+        let chain_store = self.state_manager.chain_store().clone();
+        let network = self.network.clone();
+        let genesis = self.genesis.clone();
+        let bad_block_cache = self.bad_blocks.clone();
+        let mem_pool = self.mpool.clone();
+        let block_delay = self.state_manager.chain_config().block_delay_secs as u64;
+
+        let future = async move {
+            loop {
+                let event = match p2p_messages.recv_async().await {
+                    Ok(event) => event,
+                    Err(why) => {
+                        debug!("Receiving event from p2p event stream failed: {why}");
+                        return Err(ChainMuxerError::P2PEventStreamReceive(why.to_string()));
+                    }
+                };
+
+                match Self::process_gossipsub_event(
+                    event,
+                    network.clone(),
+                    chain_store.clone(),
+                    bad_block_cache.clone(),
+                    mem_pool.clone(),
+                    genesis.clone(),
+                    PubsubMessageProcessingStrategy::DoNotProcess,
+                    block_delay,
+                )
+                .await
+                {
+                    Ok(_) => {}
+                    Err(why) => {
+                        debug!("Processing GossipSub event failed: {why:?}");
+                    }
+                };
+            }
+        };
+
+        Box::pin(future)
     }
 
     fn evaluate_network_head(&self) -> ChainMuxerFuture<NetworkHeadEvaluation, ChainMuxerError> {
@@ -828,6 +873,8 @@ enum ChainMuxerState {
     Connect(ChainMuxerFuture<NetworkHeadEvaluation, ChainMuxerError>),
     Bootstrap(ChainMuxerFuture<(), ChainMuxerError>),
     Follow(ChainMuxerFuture<(), ChainMuxerError>),
+    /// In stateless mode, forest still connects to the P2P swarm but does not sync to HEAD.
+    Stateless(ChainMuxerFuture<(), ChainMuxerError>),
 }
 
 impl<DB, M> Future for ChainMuxer<DB, M>
@@ -841,7 +888,10 @@ where
         loop {
             match self.state {
                 ChainMuxerState::Idle => {
-                    if self.sync_config.tipset_sample_size == 0 {
+                    if self.stateless_mode {
+                        info!("Running chain muxer in stateless mode...");
+                        self.state = ChainMuxerState::Stateless(self.stateless_node());
+                    } else if self.sync_config.tipset_sample_size == 0 {
                         // A standalone node might use this option to not be stuck waiting for P2P
                         // messages.
                         info!("Skip evaluating network head, assume in-sync.");
@@ -850,6 +900,11 @@ where
                         // Create the connect future and set the state to connect
                         info!("Evaluating network head...");
                         self.state = ChainMuxerState::Connect(self.evaluate_network_head());
+                    }
+                }
+                ChainMuxerState::Stateless(ref mut future) => {
+                    if let Err(why) = std::task::ready!(future.as_mut().poll(cx)) {
+                        return Poll::Ready(why);
                     }
                 }
                 ChainMuxerState::Connect(ref mut connect) => match connect.as_mut().poll(cx) {