+++ CHANGELOG.md
@@ -68,6 +68,9 @@
 
 - [#3422](https://github.com/ChainSafe/forest/issues/3422) Add NV21 (Watermelon)
   support for calibration network.
+- [#3593](https://github.com/ChainSafe/forest/pull/3593): Add `--stateless` flag
+  to `forest`. In stateless mode, forest connects to the P2P network but does
+  not sync to HEAD.
 
 ### Changed